fixed a few bugs and added some features
fixed the conversion to scientific notation sometimes not working if the number is negative. also, added a feature that will automatically add 0 if your input starts with "." or "-.". example: "." becomes "0." finally, when inputting an answer in one of the test modes, it will put a symbol on the select box, for that question, so that you know you have already answered it